requirements.txt
streamlit
pandas
langchain
google-generativeai
python-dotenv
langchain-community
langchain-google-genai